cpanel/models/requirements.txt
keras~=2.4.3
librosa~=0.8.0
numpy~=1.19.5
tensorflow~=2.5.0rc1
matplotlib~=3.4.1
pydub~=0.25.1
pandas~=1.2.4
seaborn~=0.11.1
scikit-learn~=0.24.1
pydot_ng
graphviz
python-dotenv~=0.17.0
torch~=1.8.1
torchaudio~=0.8.1